Arreglo funcion de calcular puntaje

+++ src/main/java/com/tallerwebi/dominio/ranking/RankingService.java
@@ -4,5 +4,5 @@ import com.tallerwebi.dominio.usuario.Usuario;
 
 public interface RankingService {
   // pide recibir un usuario y el puntaje base del juego actual
-  Double calcularPuntaje(Usuario usuario, Integer puntajeBase);
+  Integer calcularPuntaje(Usuario usuario, Integer puntajeBase);
 }
